refactor(link): better link with variants
