+++ apps/frontend/src/shared/ui/AppLink.vue
@@ -1,22 +1,39 @@
 <script setup lang="ts">
-import { computed } from "vue";
-import type { RouterLinkProps } from "vue-router";
+import type { ExtendedLinkProps } from "./ExtendedLink.vue";
+import ExtendedLink from "./ExtendedLink.vue";
 
-export type AppLinkProps = RouterLinkProps;
-const props = defineProps<AppLinkProps>();
+const variantClasses = {
+  primary: "text-terracotta hover:text-terracotta-dark",
+  secondary: "text-stone-600 hover:text-terracotta-500 font-medium",
+};
 
-const isExternal = computed(() => {
-  return typeof props.to === "string" && props.to.startsWith("http");
-});
+const {
+  variant = "primary",
+  icon,
+  iconPos = "right",
+} = defineProps<
+  ExtendedLinkProps & {
+    variant?: keyof typeof variantClasses;
+    icon?: string | null;
+    iconPos?: "left" | "right";
+  }
+>();
 </script>
 
 <template>
-  <RouterLink v-if="!isExternal" v-bind="$props">
-    <slot />
-  </RouterLink>
+  <ExtendedLink
+    v-bind="$props"
+    class="inline-flex items-center gap-2 text-sm transition-colors"
+    :class="variantClasses[variant]"
+  >
+    <template v-if="icon != null && iconPos === 'left'">
+      <i :class="icon" />
+    </template>
 
-  <!-- biome-ignore lint/a11y/useAnchorContent: accesibile content provided by slot or attributes -->
-  <a v-else :href="to.toString()" target="_blank" rel="noopener noreferrer">
     <slot />
-  </a>
+
+    <template v-if="icon != null && iconPos === 'right'">
+      <i :class="icon" />
+    </template>
+  </ExtendedLink>
 </template>
